requirements.txt
pip==21.3.1
setuptools==60.2.0
wheel==0.37.1
pymongo==3.10.1
python-dotenv==0.13.0

